Backend/PalindromeWorker/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub
google-cloud-storage>=1.31.0
pandas
gcsfs
requests